requirements.txt
anyio==4.4.0
APScheduler==3.10.4
asgiref==3
black==24.8.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.23.2
Werkzeug==3.0.4
